LRDOCS-11747 rename DelectableBonsai workspace to Clarity
