+++ workspaces/liferay-clarity-workspace/.gitignore
