+++ workspaces/liferay-clarity-workspace/client-extensions/liferay-clarity-batch/client-extension.yaml
@@ -1,0 +1,15 @@
+assemble:
+    - from: batch
+      into: batch
+liferay-clarity-batch:
+    name: Liferay Clarity Batch
+    oAuthApplicationHeadlessServer: liferay-clarity-batch-oauth-application-headless-server
+    type: batch
+liferay-clarity-batch-oauth-application-headless-server:
+    .serviceAddress: localhost:8080
+    .serviceScheme: http
+    name: Liferay Clarity OAuth Application Headless Server
+    scopes:
+        - Liferay.Headless.Admin.Workflow.everything
+        - Liferay.Headless.Batch.Engine.everything
+    type: oAuthApplicationHeadlessServer
