+++ workspaces/liferay-clarity-workspace/client-extensions/liferay-clarity-etc-spring-boot/LCP.json
