+++ workspaces/liferay-clarity-workspace/client-extensions/liferay-clarity-etc-spring-boot/src/main/java/com/liferay/clarity/ObjectActionAccountRestController.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.delectable.bonsai;
+package com.liferay.clarity;
 
 import com.liferay.petra.string.StringBundler;
