+++ workspaces/liferay-clarity-workspace/client-extensions/liferay-clarity-etc-spring-boot/src/main/resources/application-default.properties
@@ -1,0 +1,12 @@
+#
+# OAuth
+#
+
+liferay.oauth.application.external.reference.codes=liferay-clarity-etc-spring-boot-oauth-application-user-agent
+liferay.oauth.urls.excludes=/ready
+
+#
+# Spring Boot
+#
+
+server.port=58081
