+++ workspaces/liferay-clarity-workspace/client-extensions/liferay-clarity-etc-spring-boot/src/main/resources/application.properties
