--- workspaces/liferay-delectablebonsai-workspace/client-extensions/liferay-delectablebonsai-batch/client-extension.yaml
@@ -1,15 +1,0 @@
-assemble:
-    - from: batch
-      into: batch
-liferay-delectablebonsai-batch:
-    name: Liferay Delectable Bonsai Batch
-    oAuthApplicationHeadlessServer: liferay-delectablebonsai-batch-oauth-application-headless-server
-    type: batch
-liferay-delectablebonsai-batch-oauth-application-headless-server:
-    .serviceAddress: localhost:8080
-    .serviceScheme: http
-    name: Liferay Delectable Bonsai OAuth Application Headless Server
-    scopes:
-        - Liferay.Headless.Admin.Workflow.everything
-        - Liferay.Headless.Batch.Engine.everything
-    type: oAuthApplicationHeadlessServer
