--- workspaces/liferay-delectablebonsai-workspace/client-extensions/liferay-delectablebonsai-etc-spring-boot/client-extension.yaml
@@ -1,20 +1,0 @@
-assemble:
-    - fromTask: bootJar
-liferay-delectablebonsai-etc-spring-boot-oauth-application-user-agent:
-    .serviceAddress: localhost:58081
-    .serviceScheme: http
-    name: Liferay Delectable Bonsai Etc Spring Boot OAuth Application User Agent
-    scopes:
-        - Liferay.Headless.Admin.User.everything
-        - Liferay.Headless.Admin.Workflow.everything
-    type: oAuthApplicationUserAgent
-liferay-delectablebonsai-etc-spring-boot-object-action-account:
-    name: Liferay Delectable Bonsai Etc Spring Boot Object Action Account
-    oAuth2ApplicationExternalReferenceCode: liferay-delectablebonsai-etc-spring-boot-oauth-application-user-agent
-    resourcePath: /object/action/account
-    type: objectAction
-liferay-delectablebonsai-etc-spring-boot-workflow-action-application:
-    name: Liferay Delectable Bonsai Etc Spring Boot Workflow Action Application
-    oAuth2ApplicationExternalReferenceCode: liferay-delectablebonsai-etc-spring-boot-oauth-application-user-agent
-    resourcePath: /workflow/action/application
-    type: workflowAction
