--- workspaces/liferay-delectablebonsai-workspace/client-extensions/liferay-delectablebonsai-etc-spring-boot/src/main/resources/application-default.properties
@@ -1,12 +1,0 @@
-#
-# OAuth
-#
-
-liferay.oauth.application.external.reference.codes=liferay-delectablebonsai-etc-spring-boot-oauth-application-user-agent
-liferay.oauth.urls.excludes=/ready
-
-#
-# Spring Boot
-#
-
-server.port=58081
